Fixes __future__ import annotations handling
Fixes how we handle types in lib.py to accomodate for changes in from `__future__ import annotations`. `annotations` causes type hints to become strings to be interpreted at run time, so the current assumption that inspect will return a `Type` is incompatible.

This commit updates how we get type annotations and updates tests for versions
with annotations (3.7-3.10) to test for it.

Fixes how we handle types in lib.py to accomodate for changes in from `__future__ improt annotations`. `annotations` causes type hints to become strings to be interpreted at run time, so the current assumption that inspect will return a `Type` is incompatible.

This commit updates how we get type annotations and updates tests for versions
with annotations (3.7-3.10) to test for it.